requirements_dev.txt
Jinja2==2.7.3
MarkupSafe==0.23
Pillow==2.8.1
argparse==1.3.0
beautifulsoup4==4.3.2
gnureadline==6.3.3
google-api-python-client==1.4.0
httplib2==0.9.1
ipdb==0.8
ipython==3.1.0
itsdangerous==0.24
linecache2==1.0.0
lxml==3.4.4
oauth2client==1.4.7
pyasn1==0.1.7
pyasn1-modules==0.0.5
python-dateutil==2.4.2
requests==2.7.0
rsa==3.1.4
simplejson==3.6.5
six==1.9.0
traceback2==1.4.0
unittest2==1.0.1
uritemplate==0.6
wsgiref==0.1.2
